pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powercaldera"
version = "0.1.0"
description = "TUI para interactuar con MITRE Caldera API"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.47.0",
    "httpx>=0.27.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
powercaldera = "powercaldera.__main__:main"

[tool.setuptools.packages.find]
include = ["powercaldera*"]

[tool.setuptools.package-data]
"powercaldera" = ["app.tcss", "templates/builtin/*.json"]
